Android: Fix the flow control to not deadlock
